BioSecurePay/backend/requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
flask-jwt-extended==4.4.4
flask-cors==3.0.10
pymongo==4.3.3
bcrypt==4.0.1
cryptography==38.0.1
python-dotenv==0.21.0
requests==2.28.1
gunicorn==20.1.0
sentry-sdk[flask]==1.14.0
pytest==7.4.0
pytest-flask==1.2.0
werkzeug==2.3.8
setuptools<81